Remove module 'sendmail.py' and move its functions to 'utils.py', there is no need of a separate module
